code/requirements.txt
numpy
pandas
opencv
albumentations
pillow
scikit-learn
scipy
pytorch
torchvision
torchaudio
tqdm
tensorboard
openpyxl